codegen: implement exporing under different names
CSL has a BigInt data type which shadows the JS native BigInt.
Since we are making extensive use of the JS BigInt, we can't have a
class named BigInt in our code.

This feature will help us implement the CSL BigInt class with a
different name and export it under a new name.