manage staging and regression tests separately
This commit refactors the serialization tests
to handle staging and regression tests separately.

TXs in the staging folder that are also available
in the regression folder are automatically
filtered out, since they will be tested anyway
when the regression suite is run.

In practice, this means that one should run the
tests twice after adding new TXs to staging
to get a stable output report.

This commit also improves the report to include
failure messages.