fix: union types with multiple tagged variants not being decoded correctly
Whenever a union contains multiple variants that are encoded as a tagged
CBOR item, the code generator needs to know which tags are expected
when decoding each variant. This is not the case when there is a single
tagged variant.

With this change, now the code generator will throw an error whenever
the valid_tags field is not provided for each tagged variant in a
union that has more than one tagged variant.